backend/requirements.txt
django
djangorestframework
psycopg2-binary 
django-cors-headers 
python-dotenv 
djangorestframework-simplejwt 
pyotp
drf-yasg 
drf-spectacular